move and update toggle element to allow element to be placed inline with toggle button
