+++ src/pages/RepoPage/BundlesTab/BundleContent/ToggleElement.spec.tsx
@@ -12,6 +12,23 @@ describe('ToggleElement', () => {
     return { user }
   }
 
+  it('renders passed toggleRowElement', () => {
+    setup()
+    render(
+      <ToggleElement
+        localStorageKey="c2"
+        showElement="Show Chart"
+        hideElement="Hide Chart"
+        toggleRowElement={<p>Toggle Row Element</p>}
+      >
+        Cool contents
+      </ToggleElement>
+    )
+
+    const toggleRowElement = screen.getByText('Toggle Row Element')
+    expect(toggleRowElement).toBeInTheDocument()
+  })
+
   describe('renders open toggle', () => {
     it('toggle controls', () => {
       render(
