--- src/pages/RepoPage/BundlesTab/BundleContent/ToggleElement/index.ts
@@ -1,1 +1,0 @@
-export { ToggleElement } from './ToggleElement'
